refactor: extract PDF source run range resolution
